% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARSfunctions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{this function creates a map by State of the cases in that State.  It takes into
account the longitud and latitud of the case.  The function produces an error
when the state is invalid}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a number representing the state to be graphed}

\item{year}{the year to be graphed}
}
\value{
returns a map showing the accidents in the state
}
\description{
this function creates a map by State of the cases in that State.  It takes into
account the longitud and latitud of the case.  The function produces an error
when the state is invalid
}
\examples{
\dontrun{fars_map_state(1, 2013)}  will produce a map of accidents for state 1 in 2013
}
