% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Reads data from FARS file for specific year & plot it's accidents in specific state on map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer value representing state number to plot}

\item{year}{An integer value representing year to plot}
}
\value{
NULL
        It will plot a graphical view of accidents happening in the provided state.
}
\description{
Reads data from FARS file for specific year & plot it's accidents in specific state on map
}
\note{
This function requires a valid integer value for year, otherwise it will result in error.
      It also requires a valid integer for representing state number which should be present in data.
}
\examples{
\dontrun{
fars_map_state(state.num = 4, year=2013)
}
}
