% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{fars_map_state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{number of the state to be mapped}

\item{year}{year to get the data from}
}
\description{
Maps the data belonging to input state in input year.
}
