% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates graph for a specific state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A number indicating the state}

\item{year}{A number indicating the year}
}
\value{
This function creates a graph of a specific state and year
}
\description{
This function receives a specific state and year, retrieving the associated data.
If the state exists, it filters the data per state and creates the related graph.
The function needs the library maps.
}
\note{
The function returns an error if the state number is invalid (not present)
}
\examples{
\dontrun{
fars_map_state(state.num=1,year=2014)
}

}
