% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS data by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Number encoding of state in FARS data. Argument can be input
as integer or string. More information including codebook can be found at
NHTSA website \url{https://www.nhtsa.gov/research-data}.}

\item{year}{Four-digit year as integer or string of the data to be plotted.}
}
\value{
NULL value
}
\description{
Geographically plots highway fatalities from FARS data, subset by \code{state.nm}
and \code{year} arguments.
}
\details{
If \code{state.num} not found in data, function stopped and "invalid STATE
number" error will be returned. If no data file exists for \code{year},
call to \code{fars_read} will return "file does not exist" error.

If there are no fatalities for described subset, returns message "no accidents
to plot"

If the given value for state.num is not a state in the dataset, a "invalid STATE"
error is raised.
}
\examples{
fars_map_state(1, 2015)

}
