% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Generates map of fatalities per state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer representing each US state in the NHFTS FARS
format file in the STATE attribute}

\item{year}{A string of length 4 representing the year in the file name}
}
\value{
A plot of the state boundary polygon with fatalities as points. If
 an invalid state number or if there are no fatalities recorded for that
 state for the input year a warning message will be returned.
}
\description{
This function produces a map of the location of fatalities per state and
 input year from a standard NHFTS FARS
 /href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
 file containing that year in the filename.
}
\examples{
fars_map_state(6, 2013)

}
