% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Create a FARS map by state and year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{An integer providing the function what state to use.}

\item{year}{An integer providing the function what year to use in the filename.}
}
\value{
This function returns a graphics map object.
}
\description{
This function will create a map of fatal accidents based on the \code{state.num}
and \code{year} arguments. An error will be thrown if an invalid state number
is provided. A message will also be printed if there are zero fatalities to
map.
}
\examples{
\dontrun{fars_map_state(1,2015)}

}
