% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot fatal accident data for a given state and year on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character string giving the state number}

\item{year}{A character string giving the year for the accident data to plot}
}
\value{
This function creates a plot but does not return an object.
}
\description{
This is a simple function that plots all fatal accidents on a state map for
accidents recorded in the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (fars) that occured within that state during a given year.
An error will occur and the function will stop if the specified state number is invalid, and
a message will be shown if there are no fatal accidents within the specified state
for the specified year.
}
\examples{
fars_map_state("1","2013")

}
