% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{This function takes a numerical vector of the state and year when the accident occurred
and shows the location where traffic accidents occurred
Data should be in .csv.bz2}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A numerical vector of a particular state in the USA.}

\item{year}{A numerical vector of a year.}
}
\value{
A plot of the map of the state with points where of the accidents occurred
}
\description{
This function takes a numerical vector of the state and year when the accident occurred
and shows the location where traffic accidents occurred
Data should be in .csv.bz2
}
\note{
If state number does not exist in file will generate an error
If no records for state exist will write message and return an a invisible(NULL)
}
\examples{
\dontrun{
fars_map_state(4 , 2014)
fars_map_state('8' , '2015')
}
}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read}}
}
