% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Draws observations as point on a map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{a state number, an integer between 1 and 56.}

\item{year}{a year, as integer.}
}
\value{
\code{NULL (empty)} returned invisibly \code{\link{invisible}}.
  As side effect, the function will draw a map of accidents/observations.
}
\description{
\code{fars_map_state} takes a State number and a year as arguments.
  Reads the corresponding file and draws a map of the acccidents using 
  longitude and latitude for that the state.
}
\details{
If the value for the \code{state.num} parameter does not correspond to a
  valid state in the dataset, the function will throw an error.
  If the state and year combination has none observations, a message is
  printed.
  All the points outside of Longitude > 900 or Latitude > 90 are nullified and
  not drawn on the map.
}
\note{
The data files must be in the current working directory (see also
  \code{\link{getwd}}) to be found.
}
\section{Warning}{

  \pkg{maps} package should be attached (\code{library(maps)}),
  or the function will throw an error because maps does not export the data
  (\code{object 'stateMapEnv'}) for the \code{\link{map}} function.
}

\examples{
\donttest{fars_map_state(1, 2013)}
\donttest{fars_map_state(48, 2015)}

}
