% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Creates a map of the state and plots the points where there is an accident.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character containing State Number}

\item{year}{A character containing a year}
}
\value{
This function does not return anything. It draws a plot having a map and the points.
}
\description{
Creates a map of the state and plots the points where there is an accident.
}
\details{
This function expects a valid state number and year. This function uses internal method
\code{\link{make_filename}} to construct a file name, \code{\link{fars_read}} for reading the
file to a dtaframe. Once the dataframe is populated, this function validates if the user passed
state number exists in the dataframe and filters the records that match the state number. The coordinates
in the dataframe are filtered and draws a graph based on the range of latitudes and longitudes and
plots the points.
}
\note{
This function fails if the state number pased is invalid

This function fails if there are no entries of the state and year passed to the function.
}
\examples{
\dontrun{
fars_map_state(1)
fars_map_state("2013")
fars_map_state(1, "1999")
}

}
