% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map FARS fatal road incidents by state}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{The unique state number}

\item{year}{The year to collect data for}
}
\value{
No value is returned. Will map the FARS data.
}
\description{
This function will collect FARS csv data for a given year
and map the data for a given state number (\code{state.num}).
If the state number is not one contained in the data, the
function will give an error: \code{invalid STATE number: state.num}
Requires the dplyr, readr, maps, stats and graphics packages to be loaded.
Example data is only available for 2013:2015.
}
\examples{
\dontrun{
  setwd("data")
  year <- 2015
  state.num <- 15
  fars_map_state(state.num, year)
}

}
