% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads data from Fatality Analysis Reporting System(FARS) file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{absolute or relative path of the FARS file}
}
\value{
A Data Frame with FARS file data loaded in memory
}
\description{
Reads data from Fatality Analysis Reporting System(FARS) file
}
\note{
This function requires a valid filename path which should exist
      and should be a valid csv, otherwise it will result in error
}
\examples{
\dontrun{
data <- fars_read(filename="data/accident_2013.csv.bz2")
head(data)
}
}
