% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Filepath of the file to be read.}
}
\value{
data in tibble format.
}
\description{
Reads .csv format data from the filepath and returns a tibble with the data.
}
