% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads a csv file and transforms it into a table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string indicating the name of the file}
}
\value{
This function returns the table with the data present in the file
}
\description{
This function reads a file having \code{filename} as file name. If the file exists,
the csv file is read and converted into a table format.
}
\note{
The function returns an error if the file does not exist
}
\examples{
fars_read(filename="accident_2014.csv.bz2")

}
