% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read}
\alias{fars_read}
\title{Load data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string of the name of the file to be read}
}
\value{
This function returns a tibble of the data file and prints the first
 10 rows to the console.
}
\description{
This function reads a the USA National Highway Traffic Safety
 Administration NHFTS Fatality Analysis Reporting System FARS
 /href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
 format data file and returns a tibble.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
