% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read & Create FARS Data Frame}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string providing the FARS filename to read}
}
\value{
This function returns a tibble data frame.
}
\description{
This function will create a tibble data.frame based on a FARS filename passed
as the \code{filename} argument.
}
\examples{
\dontrun{fars_read("accident_2013.csv.bz2")}

}
