% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data File}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the file to read.}
}
\value{
A tibble (\code{tbl_df}) containing the FARS data.
}
\description{
Reads a CSV file containing Fatality Analysis Reporting System (FARS) data.
This function checks for the existence of the file, reads it using \code{readr::read_csv()},
and returns the data as a tibble.
}
\examples{
\dontrun{
data_2013 <- fars_read("accident_2013.csv.bz2")
}

}
