% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{fars_read}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a string}
}
\value{
a tibble
}
\description{
Reads a csv file and converts it to a data frame.
The file must be saved in the same directory as the function script.
}
\examples{
fars_read("accident_2013.csv.bz2")

}
