% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a data file and store the results in a data frame tbl}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the data file to read}
}
\value{
This function returns a data frame tbl (tbl_df) which contains the data read from the file "filename".
}
\description{
This is a simple function that reads a data file named "filename" into a dataframe
using the read_csv function in the readr package. The function stops if the data file
does not exist.
}
\examples{
\dontrun{fars_data_table <- fars_read("data/accident_2013.csv.bz2")}

}
