% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{This function use the read_csv function in the package readr.
It reads a single csv file into R for analysis.
the packages readr and dplyr should be installed.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string denoting the .csv or .csv.bz2 file to be read into R.}
}
\value{
An R object of class data.frame
}
\description{
This function use the read_csv function in the package readr.
It reads a single csv file into R for analysis.
the packages readr and dplyr should be installed.
}
\note{
If the file does not exist at the path it will create an error
}
\examples{
\dontrun{
path <- "accident_2013.csv.bz"
fars_read(path)
fars_read("accident_2013.csv.bz")
}
}
