% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{CSV data import}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string containing the path and the file name of a csv.-file.}
}
\value{
This function returns the imported file as a tibble data frame.
}
\description{
This function takes a file name a input to read .csv-files and imports the file in case it is available and returns a data frame as a tibble.
Messages caused by the import are suppressed and the progress bar is turned off. In case the file does not exist, the program stops and prints
out that the file does not exist.
}
\note{
The function stops if the file name provided does not exist.
}
\examples{
\dontrun{
fars_read("accident_2015.csv.bz2")
}

}
