% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character containing the file name}
}
\value{
This function returns a dataframe containing the data from the csv file passed by filename
}
\description{
Read file
}
\details{
This is a simple function that expects a csv filename as argument, checks if the file exists and if
it exists it reads the file to an object and converts it to a data frame.
}
\note{
Note that this function fails if the filename is not passed as argument.
}
\examples{
fars_read("data/accident_2013.csv.bz2")
fars_read("data/accident_2015.csv.bz2")

\dontrun{
fars_read()
fars_read("accident_2015.csv.bz2")
}

}
