% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read csv files into a data.frame.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{Path to a csv file.}
}
\value{
Returns an object of the class data.frame.
}
\description{
This function takes a csv file path and reads the file
using \code{readr::read_csv()} and \code{dplyr::tbl_df()}.
Example data is only available for 2013-2015.
Requires readr and dplyr to run.
}
\examples{
\dontrun{
  df <- fars_read("data/accident_2013.csv.bz2")
  print(class(df))
  head(df)
}

}
