% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Reads the data from input years}
}
\value{
A list with the data from input years, and an extra column with year
}
\description{
fars_read_years
}
\examples{
fars_read_years(c('1990', '1991', '2010'))

}
