% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creates a list of tables corrisponding to different years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
This function returns a list of tables corrisponding to a list of years
}
\description{
This function receives a list of years as input and retrieves the associated files.
For each file it retrieves the corrisponding data table and selects the month and year columns.
}
\note{
The function returns an error if the year is invalid:
the corrispondent file does not exist
}
\examples{
\dontrun
{
  fars_read_years(years=c(2014,2015))
}

}
