% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple data files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{vector of four-digit years as integer or strings.}
}
\value{
data.frame
}
\description{
Reads and binds into single data.frame all data files corresponding to values
in \code{years}. Dataframe reduced to variables \code{MONTH} and\code{year} before
being returned. Serves as pre-processing helper function for
\code{\link{fars_summarize_years}}.
}
\details{
If there is not a data file for any given year, "invalid year" warning will
be raised.
}
