% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Creates a tibble from a year}
\usage{
fars_read_years(years)
}
\arguments{
\item{year}{A string of length 4 giving the year, or list of years, in the
NHFTS FARS filename from which the month and year attributes will be
summarised.}
}
\value{
A tibble containing the MONTH and year attribute for each input year.
 If a year is entered that does not have the corresponding filename a
 warning message will be printed to the console.
}
\description{
This function reads in a year or a list of years and creates a tibble
 containing month and year from a standard NHFTS FARS
 /href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
 file containing that year in the filename.
}
\examples{
fars_read_years(2013)
fars_read_years(c(2013, 2014, 2015))

}
