% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{fars_read_years}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of years such as c(2013,2014,2015)
#@param year year function applied to each year in the list}
}
\value{
tibbles n = n years
}
\description{
Uses the fars_read and make_filename functions to read a list of inputted years and
applies the year function over each element of the vector,
If year is valid calls fars_read function to read the csv file associated with each year
Uses the dplyr function mutate to select month and year from data.
}
\details{
An error message is generated if an invalid year is entered in the supplied list
}
\examples{
fars_read_years(c(2013,2014,2015))
}
