% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple years of accident data from the fars database into a list}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of character strings giving the desired years to read from the fars database}
}
\value{
This function returns a list with the desired accident data for the specified list of years.
}
\description{
This is a simple function that takes a list of years as character string input
and reads the approprite accident data files from the US National Highway Traffic
Safety Administration's Fatality Analysis Reporting System (fars) for the specified years.
The results are returned in a list.  If data for a requested year is not available,
a warning is produced and nothing is returned for the given year.
}
\examples{
\dontrun{accidentData <- fars_read_years(c("2013","2015"))}

}
