% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{This function takes a numerical vector of years to be included in the data frame output.
Data should be in .csv.bz2.
If a year is not included in the data the function shows a warning and return null for that year.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A numerical vector indicating the years to include in the search}
}
\value{
A list of data frames for each year that was requested.
}
\description{
This function takes a numerical vector of years to be included in the data frame output.
Data should be in .csv.bz2.
If a year is not included in the data the function shows a warning and return null for that year.
}
\note{
If file does not exist for corresponding year a warning will be generated and NULL returned for that year
}
\examples{
\dontrun{
fars_read_years(2013)
fars_read_years(2013L)
fars_read_years('2013')
fars_read_years(c(2013, 2014))
fars_read_years(list(2013, 2015))
}
}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read_years}}
}
