% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read list of files and returns a dataframe containing all months and years from the files.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A list of characters containing the years}
}
\value{
This function returns a dataframe having the list of month and years from the files read based
on the list of years passed as parameter to function. This function also prints a warning if there is any error
while reading the file or processing the data frame.
}
\description{
Read list of files and returns a dataframe containing all months and years from the files.
}
\details{
This function that expects a list of years, for each year in the list, construct the file name.
Once the file name is created, read the content of file to the dataframe, create new field year
and return month and year fields from dataframe.
}
\note{
Note that this function fails if the years is not passed as argument.
}
\examples{
fars_read_years(c("2013", "2014"))
fars_read_years(c(2013, 2015))
fars_read_years(2013)

\dontrun{
fars_read_years()
fars_read_years("1900")
}


}
