% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Reads data from FARS file for different years to create summarized cleaner data view}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector with different year values}
}
\value{
A Data Frame representing number of fatal injuries spread across different years and corresponding months.
        Data frame consists of \code{MONTH} column with provided \code{years} columns
        where rows represents the count of fatal injuries occurring in different months
}
\description{
Reads data from FARS file for different years to create summarized cleaner data view
}
\note{
This function requires a vector with valid year representing values, otherwise it will result in error.
      Also, it will report warning if data is required for year whose file does not exists.
}
\examples{
\dontrun{
summarized_data <- fars_summarize_years(c(2013,2014,2015))
head(summarized_data)
}
}
