% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Vector with the years to be summarized}
}
\value{
Summarized table of the input years
}
\description{
fars_summarize_years
}
\examples{
\dontrun{
fars_summarize_years(c('1990', '1991', '2010'))
}
}
