% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizes results from multiple years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of years}
}
\value{
This function returns a table with summarized results for different years
}
\description{
This function receives a list of years as input and retrieves the associated data tables
with \code{fars_read_years}. Then it summarizes the results per year.
}
\examples{
\dontrun
{
  fars_summarize_years(years=c(2014,2015))
}

}
