% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarises the number of fatalities per month for a given year}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A string of length 4 giving the year, or list of years, in the
NHFTS FARS filename from which the month and year attributes will be
summarised.}
}
\value{
A tibble containing the number of fatalities per month for each
 input year.
}
\description{
This function reads in a year or a list of years and creates a tibble
 summarising the number of fatalities per month as rows and year as
 columns from a standard NHFTS FARS
 /href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
 file containing that year in the filename.
}
\examples{
fars_summarize_years(2013)
fars_summarize_years(c(2013, 2014, 2015))

}
