% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Create a summary of FARS data set by month and year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A vector or list of years to select.}
}
\value{
This function returns a summary tibble data frame of the accidents
per month.
}
\description{
This function will return a summary of total number of fatalities per month
and year. The \code{years} argument can be a vector or list
of year(s) in integer format.
}
\examples{
\dontrun{fars_summarize_years(2015)}

}
