% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{fars_summarize_years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{(c(2013,2014,2015))}
}
\value{
a tibble
}
\description{
Uses helper functions years and fars_read_years function to read each csv file. the dplyr bind_rows function
combines the data frames into dat_list object then groups by year and month.
The year and number of accidents for each month are spread to columns.
}
\examples{
fars_summarize_years(c(2013,2014,2015))

}
