% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{This functions creates a data frame containing number of records available per month}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A numerical vector indicating which years to include in the output dataframe.}
}
\value{
A data frame with the variables MONTH and each year where data is
  avaiable, summarizing the number of observations recorded each month per year.
}
\description{
This functions creates a data frame containing number of records available per month
}
\note{
If file does not exist for corresponding year a warning will be generated.
}
\examples{
\dontrun{
fars_summarize_years(2013)
fars_summarize_years(2013L)
fars_summarize_years('2013')
fars_summarize_years(c(2013, 2014))
fars_summarize_years(list(2013, 2015))
}
}
\seealso{
\code{\link{make_filename}}

\code{\link{fars_read_years}}
}
