% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarizing the amount of records per Month and year for a given input vector of years}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A year or a vector of years for which you assume that FARS data is available, e.g. "2015", c(2014:2015), 2014:2016}
}
\value{
This function returns the amount of records for each month and year from the FARS data given the argument year.
}
\description{
This function takes a year or a vector of years as input and gives back a data frame showing the amount of records for each month and each year
for each input year from the FARS data.
}
\note{
The function brings up a warning in case the data does not exist for the year provided to the function.
}
\examples{
fars_summarize_years(2014)
fars_summarize_years(2014:2016)

}
