% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Read list of files and returns a dataframe containing list of months and list of records for each year.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A list of characters containing the years}
}
\value{
This function returns a dataframe having the summary with list of months and number of records
in the files for each year. This function also prints a warning if there is any error
while reading the files or processing the data frame.
}
\description{
Read list of files and returns a dataframe containing list of months and list of records for each year.
}
\details{
This function expects a list of years, reads the content from all the files of these years
using the internal method \code{\link{fars_read_years}}  for each year in the list and saves the result
to a data frame. This dataframe is then processed to get a a final dataframe having the summary with
list of months and total number of records for each year.
}
\note{
Note that this function fails if the years is not passed as argument.
}
\examples{
fars_summarize_years(c("2013", "2014"))
fars_summarize_years("2014")

\dontrun{
fars_summarize_years()
fars_summarize_years("1900")
}

}
