% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates a valid FARS filename as per the provided year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer value for providing year to create corresponding filename}
}
\value{
A character vector with required valid FARS filename for provided year
}
\description{
Creates a valid FARS filename as per the provided year
}
\note{
This function requires a valid integer value for year, otherwise it will result in error.
      For example, this code \code{make_filename("abc")} will break.
}
\examples{
\dontrun{
fars_filename <- make_filename(2013)
print(fars_filename)
}
}
