% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{year to be pasted in the output name}
}
\value{
Valid file name with the input year.
}
\description{
Creates a valid file name for csv.bz2 format with accident prefix using the year introduced as an input.
}
\examples{

make_filename('1990')

}
