% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates the file name starting from the year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number indicating the year}
}
\value{
This function returns the string with the file name associated with the year
}
\description{
This function creates the file name giving the \code{year} as input.
}
\examples{
\dontrun
{
  make_filename(year=2014)
}

}
