% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An four-digit year to be assigned to
file name. Function accepts both integers and strings as it converts the
latter to an integer before assembling file name.}
}
\value{
Returns a string in FARS format with \code{year} embedded. The file
name has .csv.bz2 file extensions.
}
\description{
Formats string file name to facilitate loading FARS files.
}
\details{
Helper function used with other functions in this package, specifically
\code{\link{fars_read_years}}, \code{\link{fars_map_state}}.
}
\examples{
make_filename("2015")
make_filename("2016")

}
\seealso{
\code{fars_read} is called as a helper function in
\code{\link{fars_read_years}}, \code{\link{fars_map_state}}
}
