% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsfunctions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename with a certain year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A string of length 4 giving the year to be added to the filename.}
}
\value{
This function returns a chatacter vector.
}
\description{
This function reads in a year and creates a standard NHFTS FARS
 /href{https://www.nhtsa.gov/research-data/fatality-analysis-reporting-system-fars}
 filename as #'  a character string vector.
}
