% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create FARS Filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer providing the function what year to use in the filename.}
}
\value{
This function returns a string of a file name compatible to other
farspackage functions.
}
\description{
This function will create a file name using the \code{year} argument that
will be compatible for use in the other farspackage functions.
}
\examples{
\dontrun{make_filename(2014)}

}
