% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates the filename for a FARS data file based on the given year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer or string representing the year.}
}
\value{
A character string representing the filename.
}
\description{
Creates the filename for a FARS data file based on the given year.
}
\examples{
make_filename(2015)

}
