% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{make_filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer value representing a year}
}
\value{
[1] "accident_2015.csv.bz2"
}
\description{
Helper function allows user to input a year which is returned as the integer parameter %d in filename.
The sprintf function returns a formatted combination of text and variable values.
}
\examples{
make_filename(2015)
}
