% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a character string for the name of a fars database accident file for a given year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character string giving the desired year for which to create the filename}
}
\value{
This function returns a character string containing the desired filename.
}
\description{
This is a simple function that creats a filename for a compressed (.bz2) comma-separated-variable (.csv)
data file for accidents recorded in the US National Highway Traffic Safety Administration's
Fatality Analysis Reporting System (fars) that occured for a given year ("year").  The created filename is of
the form: "accident_year.csv.bz2" where "year" is replaced with the desired year.  There are no
conditions that will cause an error.
}
\examples{
file2013 <- make_filename("2013")

}
