% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{This function use the sprintf function to create a filename
for the data in the year  observed in csv.bz2 format.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A numerical vector for the year to include in the file name}
}
\value{
A character vector indicating the name of the files in .csv.bz2 format
}
\description{
This function use the sprintf function to create a filename
for the data in the year  observed in csv.bz2 format.
}
\note{
If input cannot be coerced to integer(s) warning will result and NA(s) will be returned.
}
\examples{
\dontrun{
make_filename(2013)
make_filename('2013')
make_filenam(c(2013,2015))
}
}
