% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create or construct filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character containing the year.}
}
\value{
This function returns a final filename having the year.
}
\description{
Create or construct filename
}
\details{
This is a simple function which expects a year as an argument, converts the year to integer, constructs the
file name by appending year to the filename and returns it.
}
\note{
Note that this function fails if argument year is not passed as argument.
}
\examples{
make_filename("2001")
make_filename(1999)

\dontrun{
make_filename()
make_filename("Test")
}


}
