% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Get Fatality-Analysis-Reporting-System (FARS) csv filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{The year for the file name you want to return.}
}
\value{
Returns a character string giving the name of a csv file.
}
\description{
This function returns the filename for FARS csv data given
a certain year. For more info, see \href{https://www.nhtsa.gov/Data/Fatality-Analysis-Reporting-System-(FARS)}{the FARS website}.
Example data is only available for 2013-2015.
}
\examples{
  print(make_filename("2015"))

}
